requirements.txt
django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
Pillow
drf-yasg
python-decouple
django-filter
django-allauth
# social-auth-app-django
# django-rest-framework-social-oauth2